move browser adapter to it's own folder
