+++ src/background/chrome-feature-controller.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { BrowserAdapter } from './browser-adapter';
+import { BrowserAdapter } from './browser-adapters/browser-adapter';
 
 export class ChromeFeatureController {
     public static configureCommandTabUrl: string = 'chrome://extensions/configureCommands';
