+++ src/background/message-distributor.ts
@@ -5,7 +5,7 @@ import { createDefaultLogger } from '../common/logging/default-logger';
 import { Logger } from '../common/logging/logger';
 import { InterpreterMessage } from '../common/message';
 import { Tab } from './../common/itab.d';
-import { BrowserAdapter } from './browser-adapter';
+import { BrowserAdapter } from './browser-adapters/browser-adapter';
 import { GlobalContext } from './global-context';
 import { TabToContextMap } from './tab-context';
