+++ src/popup/popup-initializer.ts
@@ -4,7 +4,7 @@ import { autobind } from '@uifabric/utilities';
 import { loadTheme } from 'office-ui-fabric-react';
 import * as ReactDOM from 'react-dom';
 
-import { BrowserAdapter } from '../background/browser-adapter';
+import { BrowserAdapter } from '../background/browser-adapters/browser-adapter';
 import { A11YSelfValidator } from '../common/a11y-self-validator';
 import { AxeInfo } from '../common/axe-info';
 import { VisualizationConfigurationFactory } from '../common/configs/visualization-configuration-factory';
