+++ src/popup/support-link-handler.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { BrowserAdapter } from '../background/browser-adapter';
+import { BrowserAdapter } from '../background/browser-adapters/browser-adapter';
 import { config } from '../common/configuration';
 import { WindowUtils } from '../common/window-utils';
