+++ src/tests/unit/tests/background/actions/inspect-action-creator.test.ts
@@ -4,7 +4,7 @@ import * as _ from 'lodash';
 import { IMock, It, Mock, MockBehavior, Times } from 'typemoq';
 import { InspectActionCreator } from '../../../../../background/actions/inspect-action-creator';
 import { InspectActions, InspectPayload } from '../../../../../background/actions/inspect-actions';
-import { BrowserAdapter, ChromeAdapter } from '../../../../../background/browser-adapter';
+import { BrowserAdapter, ChromeAdapter } from '../../../../../background/browser-adapters/browser-adapter';
 import { InspectMode } from '../../../../../background/inspect-modes';
 import { TelemetryEventHandler } from '../../../../../background/telemetry/telemetry-event-handler';
 import { Action } from '../../../../../common/flux/action';
