+++ src/tests/unit/tests/background/actions/tab-action-creator.test.ts
@@ -5,7 +5,7 @@ import { IMock, It, Mock, MockBehavior, Times } from 'typemoq';
 import { PageVisibilityChangeTabPayload, SwitchToTargetTabPayload } from '../../../../../background/actions/action-payloads';
 import { TabActionCreator } from '../../../../../background/actions/tab-action-creator';
 import { TabActions } from '../../../../../background/actions/tab-actions';
-import { BrowserAdapter, ChromeAdapter } from '../../../../../background/browser-adapter';
+import { BrowserAdapter, ChromeAdapter } from '../../../../../background/browser-adapters/browser-adapter';
 import { TelemetryEventHandler } from '../../../../../background/telemetry/telemetry-event-handler';
 import { Action } from '../../../../../common/flux/action';
 import { Tab } from '../../../../../common/itab';
