+++ src/tests/unit/tests/background/stores/assessment-store.test.ts
@@ -20,7 +20,7 @@ import {
 import { AssessmentActions } from '../../../../../background/actions/assessment-actions';
 import { AssessmentDataConverter } from '../../../../../background/assessment-data-converter';
 import { AssessmentDataRemover } from '../../../../../background/assessment-data-remover';
-import { ChromeAdapter } from '../../../../../background/browser-adapter';
+import { ChromeAdapter } from '../../../../../background/browser-adapters/browser-adapter';
 import { InitialAssessmentStoreDataGenerator } from '../../../../../background/initial-assessment-store-data-generator';
 import { AssessmentStore } from '../../../../../background/stores/assessment-store';
 import { AssesssmentVisualizationConfiguration } from '../../../../../common/configs/visualization-configuration-factory';
