+++ src/tests/unit/tests/background/telemetry/application-telemetry-data-factory.test.ts
@@ -3,7 +3,7 @@
 import { IMock, Mock, Times } from 'typemoq';
 
 import { ApplicationBuildGenerator } from '../../../../../background/application-build-generator';
-import { ChromeAdapter } from '../../../../../background/browser-adapter';
+import { ChromeAdapter } from '../../../../../background/browser-adapters/browser-adapter';
 import { InstallDataGenerator } from '../../../../../background/install-data-generator';
 import {
     ApplicationTelemetryData,
