+++ src/tests/unit/tests/issue-filing/common/issue-filing-controller-impl.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { It, Mock, Times } from 'typemoq';
 
-import { BrowserAdapter } from '../../../../../background/browser-adapter';
+import { BrowserAdapter } from '../../../../../background/browser-adapters/browser-adapter';
 import { BaseStore } from '../../../../../common/base-store';
 import { EnvironmentInfo } from '../../../../../common/environment-info-provider';
 import { CreateIssueDetailsTextData } from '../../../../../common/types/create-issue-details-text-data';
