+++ src/views/insights/dependencies.ts
@@ -2,7 +2,8 @@
 // Licensed under the MIT License.
 import { loadTheme } from 'office-ui-fabric-react';
 import * as ReactDOM from 'react-dom';
-import { ChromeAdapter } from '../../background/browser-adapter';
+
+import { ChromeAdapter } from '../../background/browser-adapters/browser-adapter';
 import { initializeFabricIcons } from '../../common/fabric-icons';
 import { ActionMessageDispatcher } from '../../common/message-creators/action-message-dispatcher';
 import { ContentActionMessageCreator } from '../../common/message-creators/content-action-message-creator';
